Extend clojure.test using multi method 'assert-expr'
Instead of having separate macro 'contains-exactly?' which replaces the
'is' macro we now extend clojure.test by adding two new assertions

(is (equal? actual expected))
AND
(is (not-equal? actual expected))

This is a de facto way that most of the other testing frameworks use for
adding custom assertions and it is a recommended approach according to
clojure.test documentation.

Now 'with-junit-output' also works as expected. It was broken in version 1.0.0.

Note! This does not provide as nice pretty printing yet than previous version did.